% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CircSpaceTime.R
\docType{package}
\name{CircSpaceTime}
\alias{CircSpaceTime}
\alias{CircSpaceTime-package}
\title{CircSpaceTime: implementation of Bayesian models,
for spatial and spatio-temporal interpolation of circular data.}
\description{
The CircSpaceTime package provides two categories of important functions:
Sampling Functions and Posterior (Kriging) Estimation Functions.
}
\section{CircSpaceTime main functions}{


\code{\link{WrapSp}} and \code{\link{ProjSp}}, for sampling
from a spatial Normal Wrapped and Projected, respectively.
\code{\link{WrapKrigSp}} and \code{\link{ProjKrigSp}}, for posterior
estimation on spatial Normal Wrapped and Projected, respectively.

\code{\link{WrapSpTi}} and \code{\link{ProjSpTi}}, for sampling
from a spatio-temporal Normal Wrapped and Projected, respectively.
\code{\link{WrapKrigSpTi}} and \code{\link{ProjKrigSpTi}}, for posterior
estimation on spatio-temporal Normal Wrapped and Projected, respectively.
}

